+++ runtime/Java/src/org/antlr/v4/runtime/CommonTokenStream.java
@@ -82,11 +82,11 @@ public class CommonTokenStream extends BufferedTokenStream {
     }
 
     @Override
-    public Token LT(int k) {
+    public Token LT(long k) {
         //System.out.println("enter LT("+k+")");
         lazyInit();
         if ( k == 0 ) return null;
-        if ( k < 0 ) return LB(-k);
+        if ( k < 0 ) return LB((int)-k);
         int i = p;
         int n = 1; // we know tokens[p] is a good one
         // find k good tokens
