+++ runtime/Java/src/org/antlr/v4/runtime/ProxyErrorListener.java
@@ -69,8 +69,8 @@ public class ProxyErrorListener implements ANTLRErrorListener {
 	@Override
 	public void reportAmbiguity(Parser recognizer,
 								DFA dfa,
-								int startIndex,
-								int stopIndex,
+								long startIndex,
+								long stopIndex,
 								boolean exact,
 								BitSet ambigAlts,
 								ATNConfigSet configs)
@@ -83,8 +83,8 @@ public class ProxyErrorListener implements ANTLRErrorListener {
 	@Override
 	public void reportAttemptingFullContext(Parser recognizer,
 											DFA dfa,
-											int startIndex,
-											int stopIndex,
+											long startIndex,
+											long stopIndex,
 											BitSet conflictingAlts,
 											ATNConfigSet configs)
 	{
@@ -96,8 +96,8 @@ public class ProxyErrorListener implements ANTLRErrorListener {
 	@Override
 	public void reportContextSensitivity(Parser recognizer,
 										 DFA dfa,
-										 int startIndex,
-										 int stopIndex,
+										 long startIndex,
+										 long stopIndex,
 										 int prediction,
 										 ATNConfigSet configs)
 	{
