+++ runtime/Java/src/org/antlr/v4/runtime/TokenFactory.java
@@ -44,7 +44,7 @@ public interface TokenFactory<Symbol extends Token> {
 	 */
 	@NotNull
 	Symbol create(@NotNull Pair<TokenSource, CharStream> source, int type, String text,
-				  int channel, int start, int stop,
+				  int channel, long start, long stop,
 				  int line, int charPositionInLine);
 
 	/** Generically useful */
